backend/requirements.txt
django>=5.1
djangorestframework>=3.14
django-cors-headers>=4.3
python-dotenv>=1.0
Pillow>=10.1.0
channels>=4.0
daphne>=4.0
channels-redis>=4.1
